Refactor: 팩토리 사용
